+++ examples/javascript/boards/updateBoard.js
@@ -1,0 +1,17 @@
+var apiKey = process.env.TRELLO_API_KEY || 'YOUR_API_KEY';
+var oauthToken = process.env.TRELLO_OAUTH_TOKEN || 'OAUTH_TOKEN';
+
+var Trello = require('../../../lib/trello-node-api')(apiKey, oauthToken);
+
+var boardRequest = function () {
+    var id = 'BOARD_ID';
+    var data = {name: 'BOARD'};
+
+    Trello.board.update(id, data).then(function (response) {
+        console.log('response ', response);
+    }).catch(function (error) {
+        console.log('error', error);
+    });
+};
+
+boardRequest();