+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.0.3 December 02, 2017
+  - Add support for TypeScript
+  - Add JavaScript examples
+  - Add TypeScript examples
+  - Add travis CI
+
 0.0.2 October 15, 2017
   - Add details to the README.MD